requirements.txt
pyyaml
apscheduler
apscheduler
bilibili_api
PyQt5
httpx
qrcode
Pillow
pyinstaller